refactor elevator to on-rio pid also
